16/34

+++ gs-serving-web-content/complete/src/main/java/com/example/servingwebcontent/repository/TenantRepository.java
@@ -3,5 +3,5 @@ package com.example.servingwebcontent.repository;
 import com.example.servingwebcontent.model.Tenant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TenantRepository extends JpaRepository<Tenant, Long> {
+public interface TenantRepository extends JpaRepository<Tenant, Integer> {
 }
